.pipelines/prchecks/CveSpecFilePRCheck/requirements.txt
openai>=1.63.0
azure-identity>=1.16.1